fix: remaining build warnings + vendor-manifest auth
Source patches (9 warnings):
- wsm.c: %d -> %zu for size_t fields (6 warnings)
- fwio.c: cast pointer arithmetic to (unsigned int) (1 warning)
- DAP_config.h: #undef guard for DAP_SWD redefinition (1 warning)
- swd_host.c: __maybe_unused on swd_get_apsel (1 warning)

Workflow fix:
- release.yml: use git clone with PAT in URL instead of gh repo clone
  for vendor-manifest push auth; add continue-on-error; suppress token
  echo in logs (set -eu not -eux)

+++ drivers/net/wireless/xr829/wlan/fwio.c
@@ -646,8 +646,8 @@ int pas_ram_check(struct xradio_common *hw_priv)
 				unsigned char *check_addr = read_buf;
 				while (check_addr < read_buf + WRITE_READ_MAX_LEN) {
 					if (*check_addr != write_value[i]) {
-						printk(KERN_ERR "[pas_ram err]:error address:0x%08x, write value:0x%02x, but read value:0x%02x\n",
-								addr + check_addr - read_buf, write_value[i], *check_addr);
+					printk(KERN_ERR "[pas_ram err]:error address:0x%08x, write value:0x%02x, but read value:0x%02x\n",
+							(unsigned int)(addr + (check_addr - read_buf)), write_value[i], *check_addr);
 					}
 					check_addr++;
 				}
